requirements.txt
Pillow==9.3.0
numpy==1.23.3
matplotlib==3.6.1
